+++ src/main/java/org/phoenix/planet/controller/CardCompanyController.java
@@ -3,7 +3,7 @@ package org.phoenix.planet.controller;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.phoenix.planet.dto.card.CardCompanyListResponse;
+import org.phoenix.planet.dto.card.response.CardCompanyListResponse;
 import org.phoenix.planet.service.card.CardCompanyService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
