+++ src/main/java/org/phoenix/planet/dto/card/response/CardCompanyListResponse.java
@@ -1,4 +1,4 @@
-package org.phoenix.planet.dto.card;
+package org.phoenix.planet.dto.card.response;
 
 public record CardCompanyListResponse(
     Long cardCompanyId,
