+++ src/main/java/org/phoenix/planet/mapper/CardCompanyMapper.java
@@ -2,7 +2,7 @@ package org.phoenix.planet.mapper;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
-import org.phoenix.planet.dto.card.CardCompanyListResponse;
+import org.phoenix.planet.dto.card.response.CardCompanyListResponse;
 
 @Mapper
 public interface CardCompanyMapper {
