+++ src/main/java/org/phoenix/planet/service/card/CardCompanyService.java
@@ -1,7 +1,7 @@
 package org.phoenix.planet.service.card;
 
 import java.util.List;
-import org.phoenix.planet.dto.card.CardCompanyListResponse;
+import org.phoenix.planet.dto.card.response.CardCompanyListResponse;
 
 public interface CardCompanyService {
