+++ src/main/java/org/phoenix/planet/service/card/CardCompanyServiceImpl.java
@@ -3,7 +3,7 @@ package org.phoenix.planet.service.card;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.phoenix.planet.dto.card.CardCompanyListResponse;
+import org.phoenix.planet.dto.card.response.CardCompanyListResponse;
 import org.phoenix.planet.mapper.CardCompanyMapper;
 import org.springframework.stereotype.Service;
