+++ src/main/resources/mapper/card/card_company.xml
@@ -6,7 +6,7 @@
 <mapper namespace="org.phoenix.planet.mapper.CardCompanyMapper">
     
     <select id="selectAll"
-        resultType="org.phoenix.planet.dto.card.CardCompanyListResponse">
+        resultType="org.phoenix.planet.dto.card.response.CardCompanyListResponse">
         SELECT
         cc.card_company_id AS cardCompanyId,
         cc.name AS name
